Refactored sidebar component
